# tp_final/src/generated/schemas.ts
// AUTO-GENERATED by npm run generate
// Do not edit manually — regenerate from openapi.yaml

import { z } from "zod";

export const CampaignSchema = z.object({ "id": z.string(), "title": z.string(), "status": z.enum(["pending","processing","completed","error"]), "createdAt": z.string().datetime({ offset: true }), "updatedAt": z.string().datetime({ offset: true }) });
export const ItemSchema = z.object({ "id": z.string(), "campaignId": z.string(), "name": z.string(), "varietals": z.array(z.string()), "producer": z.string(), "region": z.string(), "vintage": z.number().int(), "pricePerUnit": z.number().int().describe("price in pesos (no decimals)"), "unitsAvailable": z.number().int(), "description": z.string().optional() });
export const OrderSchema = z.object({ "id": z.string(), "campaignId": z.string(), "buyerPhone": z.string(), "buyerName": z.string(), "subtotalPesos": z.number().int(), "status": z.enum(["open","closed","paid"]), "createdAt": z.string().datetime({ offset: true }), "updatedAt": z.string().datetime({ offset: true }) });
export const OrderLineSchema = z.object({ "id": z.string(), "orderId": z.string(), "itemId": z.string(), "quantity": z.number().int(), "lineTotalPesos": z.number().int() });
export const RollupEntrySchema = z.object({ "itemId": z.string(), "name": z.string(), "totalUnits": z.number().int(), "totalBoxes": z.number().describe("calculated as totalUnits / unitsPerBox") });
export const DistributionEntrySchema = z.object({ "buyerName": z.string(), "buyerPhone": z.string(), "subtotalPesos": z.number().int(), "paid": z.boolean() });
export const WebhookBodySchema = z.object({ "message": z.string().min(1) }).strict();
export const CreateOrderBodySchema = z.object({ "campaignId": z.number().int(), "userName": z.string().min(1), "lines": z.array(z.object({ "itemId": z.number().int(), "qty": z.number().int().min(0) })) });
export const PaymentBodySchema = z.object({ "paid": z.boolean() });
export const ErrorResponseSchema = z.object({ "error": z.string(), "statusCode": z.number().int() });